# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta API Json Rest actua como proxy SOAP a los servicios web de ARCA.",
        "title": "API proxy a los webservices de ARCA",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Sebastian Hogas",
            "email": "sehogas@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/fe/FECAEARegInformativo": {
            "post": {
                "description": "Este método permite informar para cada CAEA otorgado, la totalidad de los comprobantes emitidos y asociados a cada CAEA",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Factura Electrónica"
                ],
                "summary": "Informar comprobantes emitidos y asociados a una CAEA",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "FeCAEARegInfReqRequest",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.FeCAEARegInfReqRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wsfe.FECAEAResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/fe/FECAEASinMovimientoConsultar": {
            "get": {
                "description": "Esta operación permite consultar mediante un CAEA, cuáles fueron los puntos de venta que fueron notificados como sin movimiento. El cliente envía el requerimiento, el cual es atendido por el WS, superadas las validaciones de seguridad se informa el CAEA, puntos de venta identificados como sin movimientos y fecha de proceso. En caso de informar el punto de venta, se informan los datos vinculados a ese punto de venta en particular.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Factura Electrónica"
                ],
                "summary": "Consulta de Puntos Venta sin movimientos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CAEA",
                        "name": "CAEA",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Punto de Venta",
                        "name": "PtoVta",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wsfe.FECotizacionResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/fe/FECAESolicitar": {
            "post": {
                "description": "Solicitar CAE",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Factura Electrónica"
                ],
                "summary": "Solicitar CAE",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "FECAESolicitarRequest",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.FECAESolicitarRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wsfe.FECAEResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/fe/FECompConsultar": {
            "get": {
                "description": "Esta operación permite consultar mediante tipo, numero de comprobante y punto de venta los datos de un comprobante ya emitido. Dentro de los datos del comprobante resultante se obtiene el tipo de emisión utilizado para generar el código de autorización.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Factura Electrónica"
                ],
                "summary": "Consulta datos de un Comprobante",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Punto de Venta",
                        "name": "PtoVta",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tipo de Comprobante",
                        "name": "CbteTipo",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Número de Comprobante",
                        "name": "CbteNro",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wsfe.FECompConsultaResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/fe/FECompTotXRequest": {
            "get": {
                "description": "Retorna la cantidad máxima de registros que se podrá incluir en un request al método FECAESolicitar / FECAEARegInformativo.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Factura Electrónica"
                ],
                "summary": "Cantidad máxima de registros por request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wsfe.FERegXReqResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/fe/FECompUltimoAutorizado": {
            "get": {
                "description": "Consultar último comprobante autorizado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Factura Electrónica"
                ],
                "summary": "Último comprobante autorizado",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Punto de venta",
                        "name": "ptoVta",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tipo de comprobante",
                        "name": "cbteTipo",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wsfe.FERecuperaLastCbteResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "dto.FECAESolicitarRequest": {
            "type": "object",
            "properties": {
                "Cabecera": {
                    "$ref": "#/definitions/wsfe.FECabRequest"
                },
                "Detalle": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wsfe.FECAEDetRequest"
                    }
                }
            }
        },
        "dto.FeCAEARegInfReqRequest": {
            "type": "object",
            "properties": {
                "Cabecera": {
                    "$ref": "#/definitions/wsfe.FECabRequest"
                },
                "Detalle": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wsfe.FECAEADetRequest"
                    }
                }
            }
        },
...
        "wsfe.ArrayOfErr": {
            "type": "object",
            "properties": {
                "Err": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wsfe.Err"
                    }
                }
            }
        },
...
        "wsfe.FECAEAResponse": {
            "type": "object",
            "properties": {
                "Errors": {
                    "$ref": "#/definitions/wsfe.ArrayOfErr"
                },
                "Events": {
                    "$ref": "#/definitions/wsfe.ArrayOfEvt"
                },
                "FeCabResp": {
                    "$ref": "#/definitions/wsfe.FECAEACabResponse"
                },
                "FeDetResp": {
                    "$ref": "#/definitions/wsfe.ArrayOfFECAEADetResponse"
                }
            }
        },
...
        "wsfe.FECAEResponse": {
            "type": "object",
            "properties": {
                "Errors": {
                    "$ref": "#/definitions/wsfe.ArrayOfErr"
                },
                "Events": {
                    "$ref": "#/definitions/wsfe.ArrayOfEvt"
                },
                "FeCabResp": {
                    "$ref": "#/definitions/wsfe.FECAECabResponse"
                },
                "FeDetResp": {
                    "$ref": "#/definitions/wsfe.ArrayOfFECAEDetResponse"
                }
            }
        },
...
        "wsfe.FECotizacionResponse": {
            "type": "object",
            "properties": {
                "Errors": {
                    "$ref": "#/definitions/wsfe.ArrayOfErr"
                },
                "Events": {
                    "$ref": "#/definitions/wsfe.ArrayOfEvt"
                },
                "ResultGet": {
                    "$ref": "#/definitions/wsfe.Cotizacion"
                }
            }
        },
...
    }
}